// src/app/layouts/dashboard/pages/enrollments/models/enrollment.ts
export interface Enrollment {
    id: string;
    courseTitle: string;
    studentName: string;
    startDate: Date;
    endDate: Date;
    shift: string;
    modality: string;
    teacher: string,
    email: string,
    magicWandCore: string,
    enrollmentId: number;
    studentId: number;
    courseId: number;
}